add button that work when robot is on the wall and doesnt use lime light

+++ src/main/java/robot/crossSubsystem/CrossSubsystemConstants.java
@@ -9,7 +9,7 @@ public class CrossSubsystemConstants {
   public static final double TIME_BETWEEN_BALLS = 100; //TODO check and  change if need
   public static final double DELAY_AFTER_SHOOT = 0; //TODO check and  change if need
   public static final double LOADER_DELAY = 0.3; //TODO check and  change if need
-  public static final int SHOOTER_VELOCITY = 39000; //TODO check and change
+  public static final double SHOOTER_VELOCITY = 39000; //TODO check and change
 
 }
